requirements.txt
pandas>=1.5
numpy>=1.20
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
PyYAML>=6.0
pytest>=7.0
pytest-mock>=3.11
great_expectations>=0.17
scikit-learn>=1.2
streamlit>=1.27
ruff>=0.4
black>=24.3
mypy>=1.7